+++ systemtest/src/main/java/io/strimzi/systemtest/utils/kafkaUtils/KafkaConnectorUtils.java
@@ -6,8 +6,8 @@ package io.strimzi.systemtest.utils.kafkaUtils;
 
 import io.strimzi.api.kafka.model.KafkaConnector;
 import io.strimzi.systemtest.Constants;
-import io.strimzi.systemtest.enums.ResourceReadiness;
 import io.strimzi.systemtest.resources.ResourceManager;
+import io.strimzi.systemtest.resources.ResourceOperation;
 import io.strimzi.systemtest.resources.crd.KafkaConnectorResource;
 import io.strimzi.test.TestUtils;
 import org.apache.logging.log4j.LogManager;
@@ -71,7 +71,7 @@ public class KafkaConnectorUtils {
     }
 
     public static void waitForConnectorCreation(String connectS2IPodName, String connectorName) {
-        long connectorTimeout = ResourceReadiness.getTimeoutForResourceReadiness(KafkaConnector.RESOURCE_KIND);
+        long connectorTimeout = ResourceOperation.getTimeoutForResourceReadiness(KafkaConnector.RESOURCE_KIND);
         TestUtils.waitFor(connectorName + " connector creation", Constants.GLOBAL_POLL_INTERVAL, connectorTimeout, () -> {
             String availableConnectors = getCreatedConnectors(connectS2IPodName);
             return availableConnectors.contains(connectorName);
